+++ mocks/controller/Handler.go
@@ -15,26 +15,12 @@ type Handler struct {
 	mock.Mock
 }
 
-// Add provides a mock function with given fields: _a0, _a1
-func (_m *Handler) Add(_a0 context.Context, _a1 runtime.Object) error {
+// Handle provides a mock function with given fields: _a0, _a1
+func (_m *Handler) Handle(_a0 context.Context, _a1 runtime.Object) error {
 	ret := _m.Called(_a0, _a1)
 
 	var r0 error
 	if rf, ok := ret.Get(0).(func(context.Context, runtime.Object) error); ok {
-		r0 = rf(_a0, _a1)
-	} else {
-		r0 = ret.Error(0)
-	}
-
-	return r0
-}
-
-// Delete provides a mock function with given fields: _a0, _a1
-func (_m *Handler) Delete(_a0 context.Context, _a1 string) error {
-	ret := _m.Called(_a0, _a1)
-
-	var r0 error
-	if rf, ok := ret.Get(0).(func(context.Context, string) error); ok {
 		r0 = rf(_a0, _a1)
 	} else {
 		r0 = ret.Error(0)